+++ src/Components/Footer/Footer.css
@@ -1,0 +1,13 @@
+.footer p {
+  margin: 0;
+}
+
+.footer {
+  background-color: #0f1119;
+  color: #fff;
+  font-size: 12px;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
