+++ src/Components/Landing Page/LandingPage.css
@@ -1,255 +1,1 @@
-.img-1,
-.img-2,
-.img-3 {
-  object-fit: cover;
-}
 
-.header-link {
-  text-decoration: none;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 300;
-  font-family: Inter, sans-serif;
-  cursor: pointer;
-}
-
-.hidden {
-  display: none;
-}
-
-.header {
-  width: 200px;
-  height: auto;
-  display: flex;
-  justify-content: space-between;
-  position: fixed;
-  top: 20px;
-  left: 20px;
-  z-index: 2;
-}
-
-.header-dark {
-  color: #171a26;
-}
-
-.features-text,
-.hero-section {
-  flex-direction: column;
-  display: flex;
-}
-
-.header-link:hover {
-  text-decoration: underline;
-}
-
-.hero-section {
-  height: 100vh;
-  background: linear-gradient(
-        to bottom,
-        rgba(0, 0, 0, 0.4) 80%,
-        rgba(0, 0, 0, 0) 100%
-      )
-      center/cover,
-    linear-gradient(to top, #171a26, rgba(255, 255, 255, 0) 20%) center/cover,
-    url('../../Images/viktor-bystrov-Gi0OMNguFaw-unsplash.jpg') center/cover;
-  align-items: center;
-  justify-content: center;
-  color: #fff;
-  text-align: center;
-  position: relative;
-}
-
-.title {
-  font-size: 48px;
-  font-weight: 800;
-  margin: 0;
-}
-
-.subtitle,
-li p {
-  font-weight: 300;
-}
-
-.subtitle {
-  font-size: 18px;
-}
-
-.down-icon {
-  color: #fff;
-  position: absolute;
-  bottom: 40px;
-  left: 50%;
-  cursor: pointer;
-  width: 20px;
-  height: auto;
-}
-
-.strava-connect-btn {
-  height: 60px;
-  width: auto;
-  margin-top: 50px;
-  cursor: pointer;
-}
-
-.faq,
-.faq-title,
-.footer p,
-ul {
-  margin: 0;
-}
-
-.features {
-  background-color: #171a26;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  color: #fff;
-}
-
-.features-images {
-  width: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.img-1,
-.img-2,
-.img-3 {
-  border-radius: 20px;
-}
-
-.img-1 {
-  width: 200px;
-  height: 155px;
-}
-
-.img-2 {
-  width: 150px;
-  height: 224px;
-}
-
-.img-3 {
-  width: 265px;
-  height: 400px;
-}
-
-.check,
-.faq {
-  height: auto;
-}
-
-.features-text {
-  align-items: flex-start;
-  padding-left: 40px;
-}
-
-.feature-heading {
-  font-size: 36px;
-  font-weight: 700;
-  margin: 0;
-  width: 375px;
-}
-
-.feature-subtitle {
-  font-size: 16px;
-  font-weight: 100;
-  width: 400px;
-}
-
-.bold-text,
-.faq-title {
-  font-weight: 700;
-}
-
-ul {
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-}
-
-.szh-accordion__item-btn,
-li {
-  align-items: center;
-  display: flex;
-}
-
-li {
-  list-style-type: none;
-  margin-bottom: -10px;
-}
-
-li p {
-  margin-left: 10px;
-  font-size: 14px;
-}
-
-.check {
-  color: #4a9df8;
-  width: 20px;
-}
-
-.faq {
-  min-height: 100vh;
-  background-color: #d9d9d9;
-  color: #36395c;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.faq-title {
-  font-size: 48px;
-}
-
-.szh-accordion {
-  width: 100vw;
-  max-width: 500px;
-  margin-top: 10px;
-}
-
-.szh-accordion__item {
-  border-bottom: 2px solid #36395c;
-}
-
-.szh-accordion__item-btn {
-  justify-content: space-between;
-  width: 100%;
-  border: none;
-  background-color: transparent;
-  padding: 0 10px;
-  cursor: pointer;
-}
-
-.szh-accordion__item-btn p {
-  font-weight: 700;
-  color: #36395c;
-  font-size: 18px;
-}
-
-.szh-accordion__item-panel {
-  padding: 0 25px 10px;
-}
-
-.szh-accordion__item-content {
-  transition: height 0.25s cubic-bezier(0, 0, 0, 1);
-}
-
-.footer {
-  background-color: #0f1119;
-  color: #fff;
-  font-size: 12px;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.strava-power-btn {
-  height: 20px;
-  width: auto;
-}
